Replace feedback implementation
